moved Ecu from canmatrix to own class, with tests
