+++ tests/test_canmatrix.py
@@ -129,34 +129,6 @@ def test_decode_signal():
     assert s8.raw2phys(3) == decimal.Decimal('0.00015')
 
 
-# BoardUnit
-def test_ecu_find_attribute():
-    ecu = canmatrix.canmatrix.Ecu(name="Gateway")
-    ecu.add_attribute("attr1", 255)
-    assert ecu.attribute("attr1") == '255'
-
-
-def test_ecu_no_attribute():
-    ecu = canmatrix.canmatrix.Ecu(name="Gateway")
-    assert ecu.attribute("wrong") is None
-    assert ecu.attribute("wrong", default=0) == 0
-
-
-def test_ecu_default_attr_from_db():
-    ecu = canmatrix.canmatrix.Ecu(name="Gateway")
-    define = canmatrix.canmatrix.Define("INT 0 255")
-    define.defaultValue = 33
-    matrix = canmatrix.canmatrix.CanMatrix(ecu_defines={"temperature": define})
-    assert ecu.attribute("temperature", db=matrix, default=2) == 33
-    assert ecu.attribute("wrong", db=matrix, default=2) == 2
-
-
-def test_ecu_repr():
-    ecu = canmatrix.canmatrix.Ecu(name="Gateway")
-    ecu.add_comment("with bug")
-    assert str(ecu) == "Ecu(name='Gateway', comment='with bug')"
-
-
 # Signal (generic functions)
 def test_signal_has_comment(some_signal):
     comment = "comment01"
